Uses Length in calc graph

+++ JHotDraw/src/main/java/org/jhotdraw/draw/PredefinedFunction.java
@@ -1,5 +1,5 @@
 package org.jhotdraw.draw;
 
 public interface PredefinedFunction {
-    double getLength();
+    int getLength();
 }
